Agent: Move MockPuppet to unit test suite
The MockPuppet is now only used by the unit tests.
