+++ monkey/tests/unit_tests/infection_monkey/master/mock_puppet.py
