+++ monkey/tests/unit_tests/infection_monkey/master/test_exploiter.py
@@ -5,10 +5,10 @@ from typing import Iterable
 from unittest.mock import MagicMock
 
 import pytest
+from tests.unit_tests.infection_monkey.master.mock_puppet import MockPuppet
 
 from infection_monkey.master import Exploiter
 from infection_monkey.model import VictimHost
-from infection_monkey.puppet.mock_puppet import MockPuppet
 
 logger = logging.getLogger()
